backend/pyproject.toml
[project]
name = "ide-backend"
version = "0.1.0"
description = "FastAPI сервис для авторизации и взаимодействия с IDE"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.115.5",
    "uvicorn[standard]==0.32.1",
    "sqlalchemy==2.0.36",
    "asyncpg==0.30.0",
    "pydantic-settings==2.7.1",
    "python-jose[cryptography]==3.3.0",
    "aiosmtplib==3.0.1",
    "email-validator==2.2.0",
    "python-multipart==0.0.17"
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"


